+++ components/formats-bsd/src/loci/formats/tiff/TiffParser.java
@@ -363,8 +363,8 @@ public class TiffParser implements Closeable {
     while (offset > 0 && offset < in.length()) {
       in.seek(offset);
       offsets.add(offset);
-      int nEntries = bigTiff ? (int) in.readLong() : in.readUnsignedShort();
-      int entryBytes = nEntries * bytesPerEntry;
+      long nEntries = bigTiff ? in.readLong() : in.readUnsignedShort();
+      long entryBytes = nEntries * bytesPerEntry;
       if (in.getFilePointer() + entryBytes + (bigTiff ? 8 : 4) > in.length()) {
         // this can easily happen when writing multiple planes to a file
         break;
